chore: Storage별 파일 분리
