+++ PawCut/PawCut/Sources/Models/Protocol/UserStorageManaging.swift
@@ -1,5 +1,5 @@
 //
-//  StorageProtocol.swift
+//  UserStorageManaging.swift
 //  PawCut
 //
 //  Created by Luminouxx on 8/21/25.
@@ -10,11 +10,4 @@ import Foundation
 protocol UserStorageManaging {
     func getIsOnboardingCompleted() -> Bool
     func setIsOnboardingCompleted(_ isCompleted: Bool)
-}
-
-protocol PetStorageManaging {
-    func getPetName() -> String
-    func setPetName(_ name: String)
-    func getPetType() -> PetType
-    func setPetType(_ type: PetType)
 }
